feat(api): move database api to admin section and resrict usage to admin
